// workhub.client/src/generate-api/models/WorkScheduleDtoPaginated.ts
/* tslint:disable */
/* eslint-disable */
/**
 * WorkHub.Server
 * No description provided (generated by Openapi Generator https://github.com/openapitools/openapi-generator)
 *
 * The version of the OpenAPI document: 1.0
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */

import { mapValues } from '../runtime';
import type { WorkScheduleDto } from './WorkScheduleDto';
import {
    WorkScheduleDtoFromJSON,
    WorkScheduleDtoFromJSONTyped,
    WorkScheduleDtoToJSON,
    WorkScheduleDtoToJSONTyped,
} from './WorkScheduleDto';

/**
 * 
 * @export
 * @interface WorkScheduleDtoPaginated
 */
export interface WorkScheduleDtoPaginated {
    /**
     * 
     * @type {Array<WorkScheduleDto>}
     * @memberof WorkScheduleDtoPaginated
     */
    data: Array<WorkScheduleDto>;
    /**
     * 
     * @type {number}
     * @memberof WorkScheduleDtoPaginated
     */
    currentPage: number;
    /**
     * 
     * @type {number}
     * @memberof WorkScheduleDtoPaginated
     */
    totalPages: number;
    /**
     * 
     * @type {number}
     * @memberof WorkScheduleDtoPaginated
     */
    totalCount: number;
    /**
     * 
     * @type {number}
     * @memberof WorkScheduleDtoPaginated
     */
    pageSize: number;
    /**
     * 
     * @type {boolean}
     * @memberof WorkScheduleDtoPaginated
     */
    readonly hasPreviousPage: boolean;
    /**
     * 
     * @type {boolean}
     * @memberof WorkScheduleDtoPaginated
     */
    readonly hasNextPage: boolean;
}

/**
 * Check if a given object implements the WorkScheduleDtoPaginated interface.
 */
export function instanceOfWorkScheduleDtoPaginated(value: object): value is WorkScheduleDtoPaginated {
    if (!('data' in value) || value['data'] === undefined) return false;
    if (!('currentPage' in value) || value['currentPage'] === undefined) return false;
    if (!('totalPages' in value) || value['totalPages'] === undefined) return false;
    if (!('totalCount' in value) || value['totalCount'] === undefined) return false;
    if (!('pageSize' in value) || value['pageSize'] === undefined) return false;
    if (!('hasPreviousPage' in value) || value['hasPreviousPage'] === undefined) return false;
    if (!('hasNextPage' in value) || value['hasNextPage'] === undefined) return false;
    return true;
}

export function WorkScheduleDtoPaginatedFromJSON(json: any): WorkScheduleDtoPaginated {
    return WorkScheduleDtoPaginatedFromJSONTyped(json, false);
}

export function WorkScheduleDtoPaginatedFromJSONTyped(json: any, ignoreDiscriminator: boolean): WorkScheduleDtoPaginated {
    if (json == null) {
        return json;
    }
    return {
        
        'data': ((json['data'] as Array<any>).map(WorkScheduleDtoFromJSON)),
        'currentPage': json['currentPage'],
        'totalPages': json['totalPages'],
        'totalCount': json['totalCount'],
        'pageSize': json['pageSize'],
        'hasPreviousPage': json['hasPreviousPage'],
        'hasNextPage': json['hasNextPage'],
    };
}

export function WorkScheduleDtoPaginatedToJSON(json: any): WorkScheduleDtoPaginated {
    return WorkScheduleDtoPaginatedToJSONTyped(json, false);
}

export function WorkScheduleDtoPaginatedToJSONTyped(value?: Omit<WorkScheduleDtoPaginated, 'hasPreviousPage'|'hasNextPage'> | null, ignoreDiscriminator: boolean = false): any {
    if (value == null) {
        return value;
    }

    return {
        
        'data': ((value['data'] as Array<any>).map(WorkScheduleDtoToJSON)),
        'currentPage': value['currentPage'],
        'totalPages': value['totalPages'],
        'totalCount': value['totalCount'],
        'pageSize': value['pageSize'],
    };
}